% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easygvtable.R
\name{easygvtable}
\alias{easygvtable}
\title{Easy GV Table Maker}
\usage{
easygvtable(inputdirectory,
            outputdirectory = tempdir())
}
\arguments{
\item{inputdirectory}{The directory containing cleaned CSV files.}

\item{outputdirectory}{The directory where you would like the Easy GV-ready
table to be written.}
}
\value{
a dataframe where each column contains every glucose value for a
given patient.
}
\description{
This function takes a directory of CGM data and makes one table with every
sensor value. For computing MAGE, quartile range, etc. These can easily be
pasted into the EasyGV spreadsheet (by Nathan R Hill. © University of Oxford
2010+)
}
\examples{
easygvtable(system.file("extdata","Cleaned_iPro_CSVs",
                        package = "cgmanalysis"))
}
