% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecsv.R
\name{makecsv}
\alias{makecsv}
\title{Convert Excel Workbooks to CSV.}
\usage{
makecsv(inputdirectory,
        outputdirectory = tempdir())
}
\arguments{
\item{inputdirectory}{The directory containing Excel files to be converted.}

\item{outputdirectory}{The directory where you would like the CSV files to be
written.}
}
\value{
a directory of CSV files with data matching the original Excel
workbooks.
}
\description{
This function iterates through a specified directory containing CGM data
stored in multiple excel files. It converts each file to a CSV file and saves
it in a new directory called "CSVs." Requires gdata package. Windows users
must have Perl installed.
}
\examples{
\dontrun{makecsv(system.file("extdata/", "iPro_Tests",
                 package = "cgmanalysis"))}
}
